complete component centralization
