+++ apps/web/src/components/ApplicationSnapshot/Sections/Skills/TechnicalAssetSkillsSnapshot.tsx
@@ -27,7 +27,7 @@ const TechnicalAssetSkillsSnapshot_Fragment = graphql(/** GraphQL */ `
 `);
 
 interface TechnicalAssetSkillsSnapshotProps {
-  query: FragmentType<typeof TechnicalAssetSkillsSnapshot_Fragment>;
+  query?: FragmentType<typeof TechnicalAssetSkillsSnapshot_Fragment>;
   experiences: SnapshotExperience[];
 }
 
@@ -38,7 +38,7 @@ const TechnicalAssetSkillsSnapshot = ({
   const application = getFragment(TechnicalAssetSkillsSnapshot_Fragment, query);
 
   const technicalAssetSkills = unpackMaybes(
-    application.pool.assetSkills
+    application?.pool.assetSkills
       ?.filter(
         (poolSkill) =>
           poolSkill?.skill?.category?.value === SkillCategory.Technical,
