+++ apps/web/src/components/ApplicationSnapshot/Sections/WorkPreferences/WorkPreferencesSnapshot.tsx
@@ -1,19 +1,11 @@
 import type { ReactElement } from "react";
 
-import type { FragmentType } from "@gc-digital-talent/graphql";
-
-import type { FlexibleWorkLocationOptions_Fragment } from "~/components/Profile/components/WorkPreferences/fragment";
-
 import type { WorkPreferencesSnapshotV1 } from "./WorkPreferencesV1";
 import WorkPreferencesV1 from "./WorkPreferencesV1";
 import { getSupportedVersionComponent } from "../utils";
 import type { SnapshotProps } from "../types";
 
-type WorkPreferencesSnapshotProps = SnapshotProps<WorkPreferencesSnapshotV1> & {
-  optionsQuery:
-    | FragmentType<typeof FlexibleWorkLocationOptions_Fragment>
-    | undefined;
-};
+type WorkPreferencesSnapshotProps = SnapshotProps<WorkPreferencesSnapshotV1>;
 
 const componentMap = {
   1: WorkPreferencesV1,
