+++ apps/web/src/pages/Applications/ApplicationSuccessPage/ApplicationSuccessPage.tsx
@@ -5,11 +5,11 @@ import { getLocale, navigationMessages } from "@gc-digital-talent/i18n";
 
 import useRoutes from "~/hooks/useRoutes";
 import type { GetPageNavInfo } from "~/types/applicationStep";
+import ApplicationSnapshot from "~/components/ApplicationSnapshot/ApplicationSnapshot";
 
 import type { ApplicationPageProps } from "../ApplicationApi";
 import ApplicationApi from "../ApplicationApi";
 import { useApplicationContext } from "../ApplicationContext";
-import ApplicationSnapshot from "~/components/ApplicationSnapshot/ApplicationSnapshot";
 
 export const getPageInfo: GetPageNavInfo = ({ application, paths, intl }) => {
   const path = paths.applicationSuccess(application.id);
