--- apps/web/src/pages/PoolCandidates/ViewPoolCandidatePage/components/ApplicationInformation/types.ts
@@ -1,14 +1,0 @@
-export const SECTION_KEY = {
-  CONTACT: "contact",
-  SCREENING: "screening",
-  GENERAL: "generalQuestions",
-  EDUCATION: "education",
-  ESSENTIAL: "essentialSkills",
-  ASSET: "assetSkills",
-  LANGUAGE: "language",
-  WORK_PREF: "workPref",
-  GOV_INFO: "govInfo",
-  DEI: "dei",
-  CITIZEN_VETERAN_PRIORITY: "citizenVeteranPriority",
-  SIGNATURE: "signature",
-} as const;
